Added custom welcome message functionality
