+++ src/main/java/leduc/command/SetWelcomeCommand.java
@@ -15,22 +15,7 @@ public class SetWelcomeCommand extends Command{
     public SetWelcomeCommand(String user){
        super(user);
     }
-    /**
-     * Returns a File object
-     * @return a file object containing the welcome message
-     */
-    public static File openFile(String filepath) throws FileException {
-        //open file, throw exception if the file doesnt exist.
-        File file;
-        file = new File(filepath);
-        try {
-            file.createNewFile();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
 
-        return file;
-    }
     /**
      * Allow to change the welcome message.
      * @param tasks leduc.task.TaskList which is the list of task.
@@ -40,7 +25,7 @@ public class SetWelcomeCommand extends Command{
     public void execute(TaskList tasks, Ui ui , Storage storage) throws FileException {
         FileWriter fileWriter = null;
         String filepath = System.getProperty("user.dir")+ "/data/welcome.txt";//get location of welcome message file
-        File file = openFile(filepath);
+        File file = Ui.openFile(filepath);
         //open fileWriter object
         try {
             fileWriter = new FileWriter(file);
